Needs AR integration to enable pass phone function
+ GameManager.bombPlanted needs to be set to true when the bomb is
planted

NOTE: All UI elements begin with a prefix that is the abbreviation of
the class’s name (without the state at the end).
This is important because the classes reference these objects by their
name!
State:
added a bool isCurrentState that checks if this state is the current
state
GameManager:
SetState(): sets and disables the isCurrentState bool.
SetMenu(): changed it to only disable the Vuforia GUI
SharedModeMenu:
PlantBomb(): initialize game variables in this funciton for now
removed initial definition of variables.