Allowed Display of Player's Name in Game Over Menu
-Had to make get functions in Player and Session classes to return the
names of players (only in shared mode). The SyncVarAttribute wouldn't
let me use getters and setters, so I just made get functions for
defuserName and planterName. Also, I made functions in Session to return
the first defuser name and first planter name. These functions would
only be used in shared mode. If it is in multiplayer mode, then only
"team 1" or "team 2" will be displayed in the game over menu (for now).
-Reverted Jin's playground prior to my changes